+++ surface/jvm/src/main/scala/wvlet/surface/package.scala
@@ -13,6 +13,8 @@
  */
 package wvlet
 
+import wvlet.surface.reflect.SurfaceFactory
+
 /**
   *
   */
